FRONTEND: look for snake cased non_field_errors in login/register api responses
The login and register views are implemented as vanilla django views
to ensure that CSRF protection is applied as recommended in the DRF
docs. This means that the camel case to snake case custom renderers
are not used to build the response and we need to check for
`non_field_errors` and not `nonFieldErrors`.